// src/app/BlockFashion/Admin/models/cotizantes.ts
export interface ICotizantes {
    codNucleo: number;
    nombNucleo: string;
    idUltRecibo: number;
    fechUltRecibo: Date;
    impUltRecibo: number;
    cantCotUltRevibo: number;
    mesCancelaUltRecibo: string;
    observaciones: string;
    cedDelegado: string;
    nombDelegado: string;
    teleDelegado: string;
    cantAfiliados: number;
    cantCotizantes: number;
}